Resolve DB test trait conflicts
